+++ R/utils-git.R
@@ -71,6 +71,25 @@ git_cfg_get <- function(name, where = c("de_facto", "local", "global")) {
   if (length(out) > 0) out else NULL
 }
 
+# more-specific case for user-name and -email
+git_user_get <- function(where = c("de_facto", "local", "global")) {
+  where <- match.arg(where)
+
+  list(
+    name = git_cfg_get("user.name", where),
+    email = git_cfg_get("user.email", where)
+  )
+}
+
+# translate from "usethis" terminology to "git" terminology
+where_from_scope <- function(scope = c("user", "project")) {
+  scope = match.arg(scope)
+
+  where_scope <- c(user = "global", project = "de_facto")
+
+  where_scope[scope]
+}
+
 # ensures that core.excludesFile is configured
 # if configured, leave well enough alone
 # if not, check for existence of one of the Usual Suspects; if found, configure
@@ -96,7 +115,6 @@ ensure_core_excludesFile <- function() {
   gert::git_config_global_set("core.excludesFile", path)
   invisible()
 }
-
 
 # Status------------------------------------------------------------------------
 git_status <- function(untracked) {