+++ NEWS.md
@@ -1,5 +1,9 @@
 # usethis (development version)
 
+* `use_tidy_eval()` is now defunct because it imports and re-exports a large
+  number of functions that are no longer needed in order to do tidy
+  evaluation (#1656).
+
 * `use_r()` and `use_test()` now work with all active files in `R/`, `src/`, 
   and `tests/testthat/` (#1566).
 
@@ -13,8 +17,9 @@
 
 * `git_sitrep()` gains two arguments: `tool` and `scope`, which enables 
   you to limit the report to, for example, `tool = "git"` or `scope = "user"`.
-  The default remains to provide a full report (@ijlyttle #1714).
-
+  The default remains to provide a full report. Also, provides more
+  feedback if git user's information is not set. (@ijlyttle, #1714, #1706).
+  
 * `use_article()` no longer adds the rmarkdown package to `Suggests`. Instead,
   if rmarkdown is not already a dependency, it's added to
   `Config/Needs/website`. This means that a package that only uses articles